lisav2: fix existing key error
The framework fails with the following error:

```
EXCEPTION: Exception calling "Add" with "2" argument(s): "Item has
already been added. Key in dictionary: 'OneVM'  Key being added:
'OneVM'"
[ERROR] Source: Line 199 in script TestControllers\TestController.psm1.
```

This error cannot be reproduced, but this fix should address
the problem, but not the cause.

Investigation is still required on why this issue appears
in the first place.